+++ packages/dash-table/tests/selenium/test_markdown_link.py
@@ -3,7 +3,7 @@ from dash_table import DataTable
 import pytest
 
 
-def get_app(cell_selectable):
+def get_app(cell_selectable, markdown_options):
     md = "[Click me](https://www.google.com)"
 
     data = [
@@ -13,7 +13,7 @@ def get_app(cell_selectable):
 
     app = dash.Dash(__name__)
 
-    app.layout = DataTable(
+    props = dict(
         id="table",
         columns=[
             dict(name="a", id="a", type="text", presentation="markdown"),
@@ -23,24 +23,43 @@ def get_app(cell_selectable):
         cell_selectable=cell_selectable,
     )
 
+    if markdown_options is not None:
+        props["markdown_options"] = markdown_options
+
+    app.layout = DataTable(**props)
+
     return app
 
 
+@pytest.mark.parametrize(
+    "markdown_options,new_tab",
+    [
+        [None, True],
+        [dict(linkTarget="_blank"), True],
+        [dict(linkTarget="_self"), False],
+    ],
+)
 @pytest.mark.parametrize("cell_selectable", [True, False])
-def test_tmdl001_copy_markdown_to_text(test, cell_selectable):
-    test.start_server(get_app(cell_selectable))
+def test_tmdl001_click_markdown_link(test, markdown_options, new_tab, cell_selectable):
+    test.start_server(get_app(cell_selectable, markdown_options))
 
     target = test.table("table")
 
     assert len(test.driver.window_handles) == 1
     target.cell(0, "a").get().find_element_by_css_selector("a").click()
-    assert target.cell(0, "a").is_selected() == cell_selectable
-    assert len(test.driver.window_handles) == 2
 
     # Make sure the new tab is what's expected
-    test.driver.switch_to_window(test.driver.window_handles[1])
-    assert test.driver.current_url.startswith("https://www.google.com")
+    if new_tab:
+        assert target.cell(0, "a").is_selected() == cell_selectable
 
-    # Make sure the cell is still selected iff cell_selectable, after switching tabs
-    test.driver.switch_to_window(test.driver.window_handles[0])
-    assert target.cell(0, "a").is_selected() == cell_selectable
+        assert len(test.driver.window_handles) == 2
+        test.driver.switch_to_window(test.driver.window_handles[1])
+        assert test.driver.current_url.startswith("https://www.google.com")
+
+        # Make sure the cell is still selected iff cell_selectable, after switching tabs
+        test.driver.switch_to_window(test.driver.window_handles[0])
+        assert target.cell(0, "a").is_selected() == cell_selectable
+
+    else:
+        assert len(test.driver.window_handles) == 1
+        assert test.driver.current_url.startswith("https://www.google.com")